as/EquDirectiveParser: Change type int to size_t for `section_table_idx`

+++ assembler/source/EquDirectiveParser.cpp
@@ -63,7 +63,7 @@ bool parsers::EquDirectiveParser::is_invalid_expression(std::string expression) 
         std::string operand = match.str();
         if (!LiteralParser::is_literal(operand))
         {
-            const int section_table_idx =
+            const size_t section_table_idx =
                 this->symbol_table_->at(operand.substr(1)).section_index;
             classification_index[section_table_idx] +=
                 (operand[0] == '-' && section_table_idx != 0 ? -1 : 1);
